+++ src/Api/BaseApi.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Devarts\PlausiblePHP\Api;
+
+use GuzzleHttp\Client;
+
+abstract class BaseApi
+{
+    protected Client $client;
+
+    public function __construct(Client $client)
+    {
+        $this->client = $client;
+    }
+}
